# src/docs/api.yaml
openapi: 3.0.0
paths:
  /edzes/ten:
    get:
      description: >-
        Lekéri a legutóbbi 10 edzés a hozzájuk tartozó gyakorlatokkal és
        izomcsoportokkal együtt
      operationId: findTen
      parameters:
        - name: userId
          required: true
          in: query
          description: A felhasználó azonosítója
          schema:
            type: number
        - name: gyakorlat
          required: true
          in: query
          description: A gyakorlat id-ja
          schema:
            type: number
        - name: isTemplate
          required: false
          in: query
          description: Sablon-e az edzés
          schema:
            type: boolean
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: A legutóbbi 10 edzés
      tags:
        - Edzes
  /edzes/intervallum:
    get:
      description: >-
        Lekér egy intervallumban lévő edzéseket a user azonosítója és, egy kezdő
        és végző dátumon, vagy kiválasztott típus intervallumon belül, a
        gyakorlatokkal és izomcsoportokkal együtt
      operationId: findManyByDate
      parameters:
        - name: page
          required: false
          in: query
          description: Oldalszám
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Elemek száma oldalanként
          schema:
            minimum: 1
            maximum: 100
            default: 10
            type: number
        - name: favoriteExercises
          required: false
          in: query
          description: Kedvenc edzés
          schema:
            default: false
            type: boolean
        - name: user_id
          required: false
          in: query
          description: A felhasználó azonosítója
          schema:
            type: number
        - name: isTemplate
          required: false
          in: query
          description: Sablon-e az edzés
          schema:
            type: boolean
        - name: type
          required: false
          in: query
          description: típus megadás, hogy milyen időintervallumot szeretnénk lekérni
          schema:
            type: week|month|halfyear|all
        - name: endDate
          required: false
          in: query
          description: Az végző dátuma
          schema:
            type: string
        - name: startDate
          required: false
          in: query
          description: A kezdő dátum
          schema:
            type: string
      responses:
        '200':
          description: Az edzések sikeresen lekérve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '404':
          description: Az edzések nem találhatóak
      security:
        - access-token: []
      summary: Egy user adott intervallumon belüli edzései
      tags:
        - Edzes
  /edzes:
    post:
      description: Létrehoz egy új edzést a megadott felhasználóhoz
      operationId: create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEdzesDto'
      responses:
        '201':
          description: Az edzés sikeresen létrehozva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '404':
          description: A felhasználó nem található
      security:
        - access-token: []
      summary: Új edzés létrehozása
      tags:
        - Edzes
    get:
      description: >-
        Lekéri az összes edzést a hozzájuk tartozó gyakorlatokkal és
        izomcsoportokkal együtt
      operationId: findAll
      parameters:
        - name: page
          required: false
          in: query
          description: Oldalszám
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Elemek száma oldalanként
          schema:
            minimum: 1
            maximum: 100
            default: 10
            type: number
        - name: favoriteExercises
          required: false
          in: query
          description: Kedvenc edzés
          schema:
            default: false
            type: boolean
        - name: user_id
          required: false
          in: query
          description: A felhasználó azonosítója
          schema:
            type: number
        - name: isTemplate
          required: false
          in: query
          description: Sablon-e az edzés
          schema:
            type: boolean
      responses:
        '200':
          description: Az edzések sikeresen lekérve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EdzesekResponseDto'
      security:
        - access-token: []
      summary: Összes edzés lekérése
      tags:
        - Edzes
  /edzes/{id}/gyakorlat/{userId}:
    post:
      description: >-
        Hozzáad egy gyakorlatot egy meglévő edzéshez és rögzíti a felhasználó
        történetében
      operationId: addGyakorlatToEdzes
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddEdzesGyakorlatDto'
      responses:
        '201':
          description: A gyakorlat sikeresen hozzáadva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '404':
          description: Az edzés vagy a gyakorlat nem található
      security:
        - access-token: []
      summary: Gyakorlat hozzáadása edzéshez
      tags:
        - Edzes
  /edzes/{id}/gyakorlat/{gyakorlatId}/{userId}:
    delete:
      description: Törli a gyakorlatot az edzésből és a felhasználó történetéből
      operationId: deleteGyakorlatFromEdzes
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
        - name: gyakorlatId
          required: true
          in: path
          description: A gyakorlat azonosítója
          schema:
            type: number
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
      responses:
        '200':
          description: ''
      security:
        - access-token: []
      summary: Gyakorlat törlése edzésből
      tags:
        - Edzes
  /edzes/{id}/gyakorlat/{gyakorlatId}/set/{userId}:
    post:
      description: >-
        Hozzáad egy szettet egy edzés gyakorlatához és rögzíti a felhasználó
        történetében
      operationId: addSetToEdzesGyakorlat
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
        - name: gyakorlatId
          required: true
          in: path
          description: A gyakorlat azonosítója
          schema:
            type: number
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddEdzesGyakorlatSetDto'
      responses:
        '201':
          description: A szett sikeresen hozzáadva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '404':
          description: Az edzés, a gyakorlat vagy a felhasználó nem található
      security:
        - access-token: []
      summary: Szett hozzáadása gyakorlathoz
      tags:
        - Edzes
  /edzes/{id}/gyakorlat/{gyakorlatId}/set/{setId}/{userId}:
    patch:
      description: Módosítja egy edzés gyakorlatának szettjét
      operationId: updateSet
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
        - name: gyakorlatId
          required: true
          in: path
          description: A gyakorlat azonosítója
          schema:
            type: number
        - name: setId
          required: true
          in: path
          description: A szett azonosítója
          schema:
            type: number
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEdzesSetDto'
      responses:
        '200':
          description: A szett sikeresen módosítva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '404':
          description: Az edzés, a gyakorlat vagy a szett nem található
      security:
        - access-token: []
      summary: Szett módosítása
      tags:
        - Edzes
    delete:
      description: Törli egy edzés gyakorlatának szettjét
      operationId: removeSet
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
        - name: gyakorlatId
          required: true
          in: path
          description: A gyakorlat azonosítója
          schema:
            type: number
        - name: setId
          required: true
          in: path
          description: A szett azonosítója
          schema:
            type: number
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
      responses:
        '200':
          description: A szett sikeresen törölve
        '404':
          description: Az edzés, a gyakorlat vagy a szett nem található
      security:
        - access-token: []
      summary: Szett törlése
      tags:
        - Edzes
  /edzes/izomcsoportok:
    get:
      description: >-
        Lekéri az összes izomcsoportot és fő izomcsoportot, ami szerepel az
        edzésekben
      operationId: getEdzesIzomcsoportok
      parameters:
        - name: user_id
          required: true
          in: query
          description: A felhasználó azonosítója
          schema:
            type: number
      responses:
        '200':
          description: Az izomcsoportok sikeresen lekérve
          content:
            application/json:
              schema:
                type: object
                properties:
                  izomcsoportok:
                    type: array
                    items:
                      type: number
                    description: Az összes érintett izomcsoport azonosítója
                  fo_izomcsoportok:
                    type: array
                    items:
                      type: number
                    description: Az összes érintett fő izomcsoport azonosítója
        '400':
          description: Hibás kérés - Érvénytelen felhasználó azonosító
      security:
        - access-token: []
      summary: Edzésekben használt izomcsoportok lekérése
      tags:
        - Edzes
  /edzes/napi:
    get:
      description: >-
        Lekér egy edzést a user azonosítója és az edzés dátuma alapján a
        gyakorlatokkal és izomcsoportokkal együtt
      operationId: findOneByDate
      parameters:
        - name: userId
          required: true
          in: query
          description: A felhasználó azonosítója
          schema:
            type: number
        - name: date
          required: true
          in: query
          description: Az edzés dátuma
          schema:
            type: string
        - name: isTemplate
          required: false
          in: query
          description: Sablon-e az edzés
          schema:
            type: boolean
      responses:
        '200':
          description: Az edzés sikeresen lekérve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '404':
          description: Az edzés nem található
      security:
        - access-token: []
      summary: Egy user adott napi edzésének részletes adatai
      tags:
        - Edzes
  /edzes/{id}:
    get:
      description: >-
        Lekér egy edzést az azonosítója alapján a gyakorlatokkal és
        izomcsoportokkal együtt
      operationId: findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
      responses:
        '200':
          description: Az edzés sikeresen lekérve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '404':
          description: Az edzés nem található
      security:
        - access-token: []
      summary: Egy edzés részletes adatai
      tags:
        - Edzes
    patch:
      description: Módosítja az edzés adatait a megadott azonosító alapján
      operationId: update
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEdzesDto'
      responses:
        '200':
          description: Az edzés sikeresen módosítva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '404':
          description: Az edzés nem található
      security:
        - access-token: []
      summary: Edzés módosítása
      tags:
        - Edzes
    delete:
      description: Törli az edzést és a hozzá tartozó gyakorlatokat
      operationId: remove
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
      responses:
        '200':
          description: Az edzés sikeresen törölve
        '404':
          description: Az edzés nem található
      security:
        - access-token: []
      summary: Edzés törlése
      tags:
        - Edzes
  /edzes/{id}/finalize/{userId}:
    patch:
      description: >-
        Módosítja az edzés véglegesítésének állapotát a megadott azonosító
        alapján
      operationId: changeEdzesFinalizedStatus
      parameters:
        - name: id
          required: true
          in: path
          description: Az edzés azonosítója
          schema:
            type: number
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
      responses:
        '200':
          description: Az edzés véglegesítésének állapota sikeresen módosítva
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '400':
          description: Hiba történt az edzés állapotának módosítása során
        '404':
          description: Az edzés nem található vagy nem tartozik a megadott felhasználóhoz
      security:
        - access-token: []
      summary: Edzés véglegesítésének állapotának módosítása
      tags:
        - Edzes
  /edzes/template/{templateId}/{userId}:
    post:
      description: Létrehoz egy új edzést egy meglévő sablon alapján
      operationId: createFromTemplate
      parameters:
        - name: templateId
          required: true
          in: path
          description: A sablon azonosítója
          schema:
            type: number
        - name: userId
          required: true
          in: path
          description: A felhasználó azonosítója
          schema:
            type: number
      responses:
        '201':
          description: Az edzés sikeresen létrehozva a sablonból
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Edzes'
        '404':
          description: A sablon vagy a felhasználó nem található
        '409':
          description: Az adott napon már létezik edzés
      security:
        - access-token: []
      summary: Új edzés létrehozása sablonból
      tags:
        - Edzes
...
info:
  title: RepVault
  description: |2-

          RepVault API
      
          A védett végpontok eléréséhez Bearer típusú JWT tokent kell használnia.
          Az authentikáció fül alatt lévő /login végponton keresztül tud tokenhez jutni. Adja meg a felhasználónevét és jelszavát,
          illetve ha még nincs felhasználója, regisztráljon az /register végponton keresztül és aztán jelentkezzen be. 
          Ezt követően kattintson az "Authorize" gombra a Swagger felületén, majd másolja be a tokenjét.
      
          Miután megadta a tokent, a Swagger automatikusan hozzáadja azt az "Authorization" fejléchez, 
          így Ön tesztelheti a védett végpontokat.
        
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:8000
    description: Development Server
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    Edzes:
      type: object
      properties:
        edzes_id:
          type: number
          description: Az edzés azonosítója
          example: 1
        edzes_neve:
          type: string
          description: Az edzés neve
          example: Melledzés
        datum:
          format: date-time
          type: string
          description: Az edzés dátuma
          example: '2024-03-15T10:00:00.000Z'
        ido:
          type: number
          description: Az edzés időtartama (percben)
          example: 60
        isFinalized:
          type: boolean
          description: Igaz ha az edzés befejeződött
          example: false
        isTemplate:
          type: boolean
          description: Igaz ha az edzés sablon
          example: false
        isFavorite:
          type: boolean
          description: Kedvenc edzés
          example: false
        gyakorlatok:
          description: Az edzéshez tartozó gyakorlatok
          type: array
          items:
            $ref: '#/components/schemas/EdzesGyakorlat'
      required:
        - edzes_id
        - edzes_neve
        - datum
        - ido
        - isFinalized
        - isTemplate
        - isFavorite
        - gyakorlatok
...